requirements.txt
python-dotenv==1.0.1
openai==1.33.0
click==8.1.7
colorama==0.4.6
groq==0.8.0
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
psutil==5.9.6
websocket-client==1.6.4
